Add files via upload
ModernFancyCV is a resume template for 2020 that aims to fulfill and surpass the expectations of any HR guy or gall. 
All important information fits on a single page while maintaining clarity.
It is aimed primarily at recent college grads, but everyone else can simply swap the "Education" and "Experience" sections and is ready to go.

I will try to add a proper README.md in time.
Also, in the future I will try to fit all important options (mainly color choices) in the .tex file and pull them out of the .cls file.

Feel free to use and customize as you see fit.